Create Custom Hook for AvailableTimePicker Component
Developed a custom hook in the `/hooks` directory to encapsulate the specific logic required for the `AvailableTimePicker` React Native component. The creation of this custom hook results in several benefits:

1. Encapsulation of Logic: By housing all the logic related to time selection and validation inside the custom hook, it streamlines the structure of the `AvailableTimePicker` component, making it more maintainable.

2. Enhanced Reusability: The custom hook can be utilized by other components that may require similar functionality, fostering code reusability within the application.

3. Separation of Concerns: This segregation aids in isolating specific functionality, making it easier to test, update, or extend in the future.

The creation of this hook aligns with modern best practices in React development and contributes to a cleaner, more modular codebase.
