+++ CtasApp/src/components/AvailableTimePicker.tsx
@@ -3,72 +3,22 @@ import {View, Text, Modal, Alert} from 'react-native';
 import RNDateTimePicker from '@react-native-community/datetimepicker';
 import AvailableTimePickerStyles from '../styles/AvailableTimePickerStyles';
 import Button from './Button';
+import useTimePicker from '../hooks/useAvailableTimePicker';
 
 const AvailableTimePicker = () => {
-  const initialTime = new Date();
-  initialTime.setHours(0, 0, 0, 0);
-  const [startTime, setStartTime] = useState(initialTime);
-  const [endTime, setEndTime] = useState(new Date(initialTime));
-  const [showStartPicker, setShowStartPicker] = useState<boolean>(false);
-  const [showEndPicker, setShowEndPicker] = useState<boolean>(false);
-  const [timeSlots, setTimeSlots] = useState<string[]>([]);
-  const [modalVisible, setModalVisible] = useState<boolean>(false);
-
-  const checkTimeValidity = () => {
-    if (startTime.getTime() === endTime.getTime()) {
-      Alert.alert(
-        'Error',
-        'The hours must be different. Please select different hours.',
-      );
-      return false;
-    }
-    return true;
-  };
-
-  const handleEndTimeChange = (event: any, selectedDate: Date | undefined) => {
-    setShowEndPicker(false);
-    if (selectedDate) {
-      setEndTime(selectedDate);
-    }
-  };
-
-  const handleStartTimeChange = (
-    event: any,
-    selectedDate: Date | undefined,
-  ) => {
-    setShowStartPicker(false);
-    if (selectedDate) {
-      setStartTime(selectedDate);
-    }
-  };
-  const generateTimeSlots = () => {
-    if (!checkTimeValidity()) {return;}
-
-    let slots = [];
-    let start = new Date(startTime);
-    let end = new Date(endTime);
-
-    if (start.getTime() > end.getTime()) {
-      end.setDate(end.getDate() + 1);
-
-    if (start.getTime() === end.getTime()) {
-      Alert.alert(
-        'Error',
-        'The hours must be different. Please select different hours.',
-      );
-      return;
-    }
-
-    while (start.getTime() !== end.getTime()) {
-      slots.push(
-        start.toLocaleTimeString([], {hour: '2-digit', minute: '2-digit'}),
-      );
-      start.setMinutes(start.getMinutes() + 30);
-    }
-
-    setTimeSlots(slots);
-    console.log('Available Time Slots:', slots);
-  };
+  const {
+    startTime,
+    endTime,
+    showStartPicker,
+    showEndPicker,
+    modalVisible,
+    handleEndTimeChange,
+    handleStartTimeChange,
+    generateTimeSlots,
+    setModalVisible,
+    setShowStartPicker,
+    setShowEndPicker,
+  } = useTimePicker();
 
   return (
     <View>
